adição de campo de inadimplencia de usuario, adição de limite de exibição na tabela de usuários e de cronogramas, adição de confirmação para excluir semanas e dias

+++ frontend/src/components/painel-adm/criar-cronograma/etapa2/CriarCronograma2.jsx
@@ -96,6 +96,8 @@ export function CriarCronograma2() {
 
     async function excluirSemana(id) {
         if (user.role != "administrador" && user.role != "distribuidor") return toast.error("Baterista não pode excluir semana");
+        const confirmacao = window.confirm("Deseja realmente remover esta semana?");
+        if (!confirmacao) return;
         try {
             const response = await api.delete(`/cronograma/${params.idCronograma}/semana/${id}`);
             setTrigger(prev => !prev);
@@ -108,6 +110,8 @@ export function CriarCronograma2() {
 
     async function excluirDia(id) {
         if (user.role != "administrador" && user.role != "distribuidor") return toast.error("Baterista não pode excluir dia");
+        const confirmacao = window.confirm("Deseja realmente remover este dia?");
+        if (!confirmacao) return;
         try {
             const response = await api.delete(`/cronograma/${params.idCronograma}/semana/${semanaId}/dia/${id}`);
             setTrigger(prev => !prev);
